+++ HISTORY.md
@@ -10,6 +10,8 @@
 - CWL: support for Cromwell runner
 - Avoid making variant_regions required during processing (by filling with
   coverage) to differentiate targeted and non analyses downstream.
+- Avoid attempts to download pre-installed S3 genomes, providing better
+  errors with missing genome installs.
 - Trimming: add explicit `polyg` option for removing 3' G stretches in NovaSeq
   and NextSeq data. Now defaults to no polyG trimming unless turned on.
 - DeepVariant and Strelka2 support for customized targeted/genome calling models